atualização firebase funcionando
